Fix: SourceController depends on method only available in certain subclasses
… without being abstract. Also, SourceController shouldn't be aware of filters
