+++ src/azul/service/app_controller.py
@@ -4,33 +4,46 @@ from typing import (
     Mapping,
 )
 
-import attr
+import attrs
 from chalice import (
     BadRequestError as BRE,
     NotFoundError,
 )
 
 from azul import (
+    CatalogName,
     R,
     RequirementError,
     config,
 )
-from azul.chalice import (
-    AppController,
+from azul.auth import (
+    Authentication,
 )
 from azul.service import (
     FileUrlFunc,
+    Filters,
     FiltersJSON,
     parse_filters,
+)
+from azul.service.source_controller import (
+    SourceController,
 )
 from azul.strings import (
     pluralize,
 )
 
 
-@attr.s(auto_attribs=True, frozen=True, kw_only=True)
-class ServiceAppController(AppController):
+@attrs.frozen(kw_only=True)
+class ServiceAppController(SourceController):
     file_url_func: FileUrlFunc
+
+    def get_filters(self,
+                    catalog: CatalogName,
+                    authentication: Authentication | None,
+                    filters: str | None = None
+                    ) -> Filters:
+        return Filters(explicit=self._parse_filters(filters),
+                       source_ids=self._list_source_ids(catalog, authentication))
 
     def _parse_filters(self, filters: str | None) -> FiltersJSON:
         try:
